+++ test/fromstdlib.jl
@@ -3,7 +3,7 @@ using MarkdownAST: MarkdownAST, Node, @ast, Document,
     Paragraph, Heading, CodeBlock, BlockQuote, DisplayMath, ThematicBreak,
     List, Item, FootnoteLink, FootnoteDefinition, Admonition,
     Table, TableHeader, TableBody, TableRow, TableCell,
-    JuliaValue, LineBreak
+    JuliaValue, LineBreak, HTMLBlock
 using Markdown: Markdown
 using Test
 
@@ -348,6 +348,14 @@ using Test
                 Link("url", "") do
                     "text"
                 end
+            end
+        end
+    end
+    @static if isdefined(Markdown, :HTMLBlock)
+        let m = Markdown.MD()
+            push!(m.content, Markdown.HTMLBlock(["<div>", "x", "</div>"]))
+            @test convert(Node, m) == @ast Document() do
+                HTMLBlock("<div>\nx\n</div>")
             end
         end
     end